pyproject.toml
[project]
name = "whisper-finetune"
description = "Script to finetune whisper with huggingface datasets."
authors = [
    { name = "Claudio Paonessa", email = "claudio.paonessa@fhnw.ch" },
    { name = "Vincenzo Timmel", email = "vincenzo.timmel@fhnw.ch" }
]
readme = "README.md"
requires-python = ">=3.11"
version = "0.1.2"
dependencies = [
    "openai-whisper==20240930",
    "pydub==0.25.1",
    "bitsandbytes==0.43.1",
    "datasets==2.17.0",
    "torchaudio==2.1.2",
    "librosa==0.10.1",
    "soundfile==0.12.1; platform_system=='Windows'",
    "tiktoken @ git+https://github.com/swiss-german-speech-to-text/tiktoken@main",
    "transformers==4.53.*",
    "wandb==0.16.2",
    "jiwer==3.0.3",
    "minlora @ git+https://github.com/cccntu/minLoRA.git@main",
    "audiomentations==0.41.0",
    "pyloudnorm==0.1.1",
    "pyroomacoustics==0.8.4",
    "numpy<2.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    # Auto-formatters
    "isort>=5.10.1",
    "black>=23.1.0",
    "flake8>=6.0.0",
    # Testing
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]
display = [
    # rich is now in main deps for colored eval metrics; keep here for extras
    "rich>=13.0.0",
]

[build-system]
requires = ["setuptools>=56.0"]
build-backend = "setuptools.build_meta"

[tool.black]
# A Python utility / library to sort Python imports.
# https://github.com/psf/black
line_length = 120

[tool.isort]
# A Python utility / library to sort Python imports.
# https://pycqa.github.io/isort/
profile = "black"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--strict-markers",
    "--tb=short",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
